+++ Realm/RLMConstants.h
@@ -22,8 +22,9 @@
 
 typedef NS_OPTIONS(NSUInteger, RLMPropertyAttributes) {
 /**
- Create an index for this property for improved search performance. Only string properties
- can be indexed. Returning this for properties of any other type will have no effect.
+ Create an index for this property for improved search performance. Only string,
+ int and bool properties can be indexed. Returning this for properties of any
+ other type will have no effect.
  */
     RLMPropertyAttributeIndexed = 1 << 2,